StyleTransfer_multi_reference/requirements.txt
numpy
openopt
cvxopt
chainer
pillow
